refactor(ui): refactor layout
